refactor: 빵집 생성/수정 로직 리팩토링
